Made numInterfaces/Gates return unsigned int Fixed signed/unsigned, wrong initialization order in ctor warnings

+++ Network/Contract/InterfaceTable.cc
@@ -107,14 +107,14 @@ void InterfaceTable::deleteInterface(InterfaceEntry *entry)
 }
 */
 
-int InterfaceTable::numInterfaceGates()
+unsigned int InterfaceTable::numInterfaceGates()
 {
     // linear search is OK because normally we have don't have many interfaces (1..4, rarely more)
     int max = -1;
     for (InterfaceVector::iterator i=interfaces.begin(); i!=interfaces.end(); ++i)
         if ((*i)->outputPort()>max)
             max = (*i)->outputPort();
-    return max+1;
+    return static_cast<unsigned int> (max+1) ;
 }
 
 InterfaceEntry *InterfaceTable::interfaceByPortNo(int portNo)
